refactor(llm): split LiteLLM model-map lookups out of llm/utils (#12974)
